- add trace to VSCode launch - comment out pushUndefined and Return - Move call frame to its own file - heading for code blocks needed to be on its own line - Show the difference between the VM starting up and a new Call frame being ran
